# feat(l2): optimize RISC0's guest program input read