--- a/crates/l2/prover/src/backend/risc0.rs
+++ b/crates/l2/prover/src/backend/risc0.rs
@@ -31,7 +31,9 @@ pub enum Error {
 
 pub fn execute(input: ProgramInput) -> Result<(), Box<dyn std::error::Error>> {
     let bytes = rkyv::to_bytes::<RkyvError>(&input)?;
-    let env = ExecutorEnv::builder().write(&bytes.as_slice())?.build()?;
+    let env = ExecutorEnv::builder()
+        .write_slice(bytes.as_slice())
+        .build()?;
 
     let executor = default_executor();