--- a/crates/l2/prover/src/guest_program/src/risc0/src/main.rs
+++ b/crates/l2/prover/src/guest_program/src/risc0/src/main.rs
@@ -1,3 +1,5 @@
+use std::io::Read;
+
 use guest_program::{execution::execution_program, input::ProgramInput};
 use risc0_zkvm::guest::env;
 use rkyv::rancor::Error;
@@ -5,7 +7,8 @@ use rkyv::rancor::Error;
 fn main() {
     println!("start reading input");
     let start = env::cycle_count();
-    let input = env::read::<Vec<u8>>();
+    let mut input = Vec::new();
+    env::stdin().read_to_end(&mut input).unwrap();
     let input = rkyv::from_bytes::<ProgramInput, Error>(&input).unwrap();
     let end = env::cycle_count();
     println!("end reading input, cycles: {}", end - start);