**Description**

RISC0's official documentation suggests using `env::stdin().read_to_end` when we need to read the input as raw bytes, because that method doesn't do (de)serialization and so does not need to copy or reinterpret the input data. This is the case because we then serialize using `rkyv`.

On the host side, `ExecutorEnvBuilder::write_slice` is used to pass in the bytes.

**Execution Times Comparison**

> Command: `cargo r -r -p ethrex-replay -F risc0 -- block --zkvm risc0 --resource cpu --action execute --rpc-url http://157.180.1.98:8545 <BLOCK>`

| Block | `main` (`1ef53cc`) |  `optimize_risc0_guest_program_input_read ` | Speedup |
| --- | --- | --- | --- |
| 23451000 | 57s | 42s | 1.37x |
| 23451100 | 36s | 25s | 1.44x |
| 23451200 | 19s | 13s | 1.46x |
| 23451300 | 43s | 30s | 1.43x |
| 23451400 | 43s | 31s | 1.39x |

**Input Reading Cycles Comparison**

> Command: `cargo r -r -p ethrex-replay -F risc0 -- block --zkvm risc0 --resource cpu --action execute --rpc-url http://157.180.1.98:8545 <BLOCK>`

| Block | `main` (`1ef53cc`) |  `optimize_risc0_guest_program_input_read ` | Diff |
| --- | --- | --- | --- |
| 23451000 | 980048869 | 21048307 | -46.56x |
| 23451100 | 755479461 | 17713531 | -42.65x |
| 23451200 | 390695293 | 9137304 | -42.76x |
| 23451300 | 865096745 | 19224750 | -45x |
| 23451400 | 832784361 | 18776551 | -44.35x |